annot: clone annotations in SdrObjects when the page is duplicated
When the page is duplicated we need to clone all SdrObjects and
with that also all the annotations contained in the SdrObjects.
The annotations must also be added to the page, so that we can
find them quickly.

Previously we only cloned the annotations and relied that the
SdrObjects would be automagically be created, but as the SdrObject
are also cloned this creates a mess with duplicated objects with
annotatins for the wrong page or annotation objects with missing
annotation.

Change-Id: I4bf53e4bd387fad9b0a8e4f43edd57c3d0fd34cc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169256
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>
Tested-by: Michael Meeks <michael.meeks@collabora.com>
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>