reduce number of SvxTextEditSourceImpl listeners on SdrModel
when we have a lot of shapes, the number of listeners attached to
SdrModel becomes a serious bottleneck because of the number of
broadcast/notify function calls.

But SvxTextEditSourceImpl can just as well listen to the SdrObject,
which achieves the same end, and reduces the bottleneck on the SdrModel.

We do have to adjust some of the broadcast sites to broadcast the
event to the SdrObject listeners as well as the SdrModel listeners.
Possibly as another change we can remove the broadcasting to the
SdrModel listeners.

This shaves 10% off the load time of a large DOCX with lots of shapes.

Noting that we do no need to listen to SdrHintKind::ModelCleared because
we will get called via SvxTextEditSourceImpl::ObjectInDestruction when
the model is destroyed.

Change-Id: I3f7f6aa2a2146e4c4f59c974aa3fed92becf5eca
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169565
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
(cherry picked from commit 4d0ba0badeb5a83d8eb5373ba5d7ef931dcf62c9)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169563
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>