Make window's default mouse position a large negative
Some operations, like DrawViewShell::Activate, trigger FuDraw::ForcePointer.
The latter may call SdrView::PickAnything to test if the pointer is over an
object. In SdrMarkView::CheckSingleSdrObjectHit, the bound rectangle of the
object is tested first, which doesn't need the object's deep processing. If
the point in in the rectangle,  the object is processed further,  which may
render the object. The hit test is performed using a tolerance.

Commit 56a1f9bea2f1b64dee4111ec6671caf42ef63c91  (sd:  extend test to check
the PDF pages aren't rendered when loaded,  2020-03-26)  made sure that the
rendering doesn't happen in the LOK mode.  In that mode, the mouse position
is unchanged from the values  set in the ImplFrameData ctor.  Previously it
was enough for the test, that the initial position was [-1, -1], which made
the point outside of the bounding rectangle of the objects. But an upcoming
fix for the map mode calculations  makes tolerance large enough to consider
the point inside the rectangle, which triggers rendering and fails the test.

This change  makes the default position  [-32767, -32737],  which should be
reliably outside of the tested rectangle.

Change-Id: Ie1142729448f7ac3c628880140f123cde66b404e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170005
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit 96c3d78f39ac4385880ae6f92f15f7ab3564ffcf)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170026
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>