lok: sd: slideshow render: some bug fixes
The suggested canvas size need to be passed to createSlideRenderer.
masterPageObjectsVisible -> masterPageObjectsVisibility as on client.
Last json message should be reset on each call to renderNextLayer.
In order to avoid to send empty layer check if the json message has
been set.
After rendering the background layer we need to return or the layer
is never received.

Change-Id: I9a83895d4f3895f3363c1b6a86fac9bd22c9fe87
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170157
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>