cool#9735 sw lok: take LOK vis area into account when deciding idle layout
E.g. paste into a paragraph that is part of an Arabic numbering in the
bugdoc (or perform any other action that invalidates all pages), the
desktop case does a synchronous layout of the visible page and does the
layout for the other pages in idle.

The LOK case is worse, there we do the entire layout synchronously,
which can result in a 274 ms hang for the bugdoc. The root of this is
that the visible area in Writer is used both to decide if invalidations
are necessary for that area (this is wanted in the LOK case) and also to
decide if the layout has to be done synchronously (this is not wanted in
the LOK case).

Fix the problem by taking the LOK visible area from SwViewShell and
work with that if possible. This reduces the time spent in the
synchronous layout to 4 ms.

This should help the interactive editing of large Writer documents in
general, but pasting into a numbering is a good example, because there
you can precisely measure if just 1 page is laid out instantly or all
pages.

(cherry picked from commit e1e77b313c9fe0fff814384a67de415e33c8b27f)

Conflicts:
	sw/qa/extras/tiledrendering2/tiledrendering2.cxx

Change-Id: Ia423079c76f8d93263a023da4bfbf2a9a5ccbf33
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171660
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>