feat: Make move/copy sheet an async jsdialog
Generally pretty standard ... I asynced this one specifically because I
wanted to be able to put it into the mobile wizard as part of
https://github.com/CollaboraOnline/online/issues/9787

I used 'mutable' for the callback here, as we modify some variables
(e.g. bDoIt) inside the callback, and I preferred to do so rather than
either shadow the variables or have separately-named things inside the
callback.

Change-Id: I23cef6330a5d5f64ac8b1746ef22516dc76f9d8e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172975
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>