cool#9992 lok doc sign: add SfxObjectShell::AfterSignContent()
Currently SfxObjectShell::CheckIsReadonly() has a hack for the LOK case
to show the signatures dialog read-only, as only that is async.

The first problem for the read-write signatures dialog is that
SfxObjectShell::ExecFile_Impl() has code after invoking
SfxObjectShell::SignDocumentContent(), which will be executed too early
if the dialog is executed async.

Fix the problem by moving the code in question into a new
SfxObjectShell::AfterSignContent(), and only invoke that as a callback
after the async run finished.

The message dialog in the moved run is still non-async, but we can deal
with that later.

(cherry picked from commit 743d08d03317855907991c29534e81c9487ea67a)

Conflicts:
	sfx2/source/doc/objserv.cxx

Change-Id: I32f0895118ac0da72105ec3a24c0294e18c05545
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172931
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>